requirements.txt

# Vector Database and Embeddings
qdrant-client==1.7.0
sentence-transformers==2.2.2

# Document Processing
PyMuPDF==1.23.14
python-docx==1.1.0
pandas==2.1.4

# LLM API
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Core Dependencies
numpy==1.24.3
uuid